Revert "static cast NDiskStubs to int"
This reverts commit 51a8ec1e89db55fefd65d8cdc360166b4c2d305f.

+++ TrackletAlgorithm/TrackBuilder.h
@@ -308,7 +308,7 @@ void TrackBuilder(
             break;
         }
       }
-      disk_stub_words: for (short j = 0 ; j < static_cast<int>(NDiskStubs); j++) {
+      disk_stub_words: for (short j = 0 ; j < NDiskStubs; j++) {
         switch (j) {
           case 0:
             diskStubWords[j][nTracks] = track.template getStubValid<NBarrelStubs>() ? track.template getDiskStubWord<NBarrelStubs>() : typename TrackFit<NBarrelStubs, NDiskStubs>::DiskStubWord(0);
